requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
google-generativeai==0.3.1
python-dotenv==1.0.0
pytest==7.4.0
requests==2.31.0
bleach==6.0.0
nltk==3.8.1
gunicorn==21.2.0